Added BabyService

+++ auth-service-api/src/main/java/by/minsk/ussr/auth/api/dto/UserProfileDto.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class UserProfileDto {
-    private long id;
+    private Integer id;
     private String userName;
     private String email;
 }
